Fix a bug where resetting the buffer didn't update the keymap parent.
This is a bit confusing because it's easy to think of the text buffer
as a data structure, not an editor mode, but it's kind of acting as
both. If it were just a data structure, it wouldn't handle key
bindings, but if it were just an editor mode, it wouldn't get replaced
when a new cart gets loaded.