Code editor: add kill, yank/paste, delete
This adds a basic form of cut and paste similar to what is found in
readline default mappings:
https://www.gnu.org/software/bash/manual/bash.html#Readline-Killing-Commands

I also wrote a more generic way to remove a character in a particular
position to handle all the edge cases when deleting a newline merges
two lines, and added a binding for "delete".

There's also now a "lastKey" property in the code editor that should
contain the last used value for the keymap. This must be assigned by
the editor that uses the code.lua api.