Code editor: smarter indentation and tab, refactor
The indentation when pressing Enter will add a level if the line ends
with "then", "do", "{", "[", "(" or a function definition.

Pressing "tab" at the beginning of a line (or in leading whitespace)
will also indent and move the cursor to the first word.
If no indentation is performed, or if "tab" is pressed a second time
it will insert a whitespace as before.

This also refactored the insertNewLine method, which was needed to fix
to reuse it in the paste method (to insert newlines without indenting).