Help: only ask input if scrolling past the screen
Currently you get the "press any key to continue" message as soon as you
reach the end of the screen, which is immediately, if the cursor is
already at the end.

With this change, the content will scroll the screen up until the point
when it wouldn't fit entirely in a single screen. Just like the "more"
program worked in DOS (except that one only asked once per entire page).

I also added a very short delay when outputing lines, for nicer feedback
, replaced some rogue tab indentations there were and added comments.