Terminal: load init from autoexec.lua if present
If an autoexec.lua file exists in the root of the disk, DiskOS will
load it during the initialization of the terminal, instead of the
welcome text.

This allows users to customize the welcome screen or add other
initialization.

The printing of the prompt was made into the method term.prompt so it
can be redefined if wanted.

Also refactored some duplicated code from term.execute