Code editor: Add functions to manipulate cursor position
This allows encapsulation of the behavior assosiated with cursor movement (like checkPos()).
Now we no longer need to check the position of the cursor every time we change it, and future behavior related to the movement can be inserted in these new functions.

Additionally, the variable "flag" has been renamed to bufferNeedsRedraw (more descriptive) and made local, so it is not necessary to call checkPos() if we do not move the cursor.
The functions gotoStartLine() and gotoEndLine() where also renamed to gotoLineStart and gotoLineEnd, respectively.